Move adding image from Utils
