+++ app/src/main/java/com/rebalance/utils/LogicUtils.kt
@@ -22,90 +22,7 @@ fun getToday(): String {
     return LocalDate.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd"))
 }
 
-fun addExpense(
-    isGroupExpense: Boolean,
-    membersSelection: SnapshotStateMap<ApplicationUser, Boolean>,
-    context: Context,
-    preferences: PreferencesData,
-    groupId: Long,
-    costValue: TextFieldValue,
-    date: MutableState<String>,
-    selectedCategory: TextFieldValue,
-    spendingName: TextFieldValue,
-    callerPhoto: Bitmap?
-) {
-    if (isGroupExpense) {
-        val activeMembers =
-            membersSelection.filterValues { flag -> flag }
-        if (activeMembers.isEmpty()) {
-            alertUser("Choose at least one member", context)
-            return
-        }
-        val resultExpense = BackendService(preferences).addExpense(
-            Expense(
-                costValue.text.toDouble(),
-                date.value.ifBlank { getToday() },
-                selectedCategory.text,
-                spendingName.text
-            ),
-            groupId
-        )
-        for (member in activeMembers) {
-            BackendService(preferences).addExpense(
-                Expense(
-                    costValue.text.toDouble() / activeMembers.size * -1,
-                    date.value.ifBlank { getToday() },
-                    selectedCategory.text,
-                    spendingName.text,
-                    resultExpense.getGlobalId()
-                ),
-                groupId
-            )
-            if (callerPhoto != null) {
-                val baos = ByteArrayOutputStream()
-                callerPhoto.compress(
-                    Bitmap.CompressFormat.PNG,
-                    100,
-                    baos
-                )
-                val b = baos.toByteArray()
-                val base64String: String = Base64.encodeToString(
-                    b,
-                    Base64.DEFAULT
-                )
 
-                BackendService(preferences).addExpenseImage(
-                    base64String,
-                    resultExpense.getGlobalId()
-                )
-            }
-        }
-    } else {
-        val resultExpense = BackendService(preferences).addExpense(
-            Expense(
-                costValue.text.toDouble(),
-                date.value.ifBlank { getToday() },
-                selectedCategory.text,
-                spendingName.text
-            ),
-            preferences.groupId
-        )
-        if (callerPhoto != null) {
-            val baos = ByteArrayOutputStream()
-            callerPhoto.compress(
-                Bitmap.CompressFormat.PNG,
-                100,
-                baos
-            )
-            val b = baos.toByteArray()
-            val base64String: String = Base64.encodeToString(
-                b,
-                Base64.DEFAULT
-            )
-            BackendService(preferences).addExpenseImage(base64String, resultExpense.getGlobalId())
-        }
-    }
-}
 
 fun createGroup(
     groupCurrency: TextFieldValue,
